frontend: tweak the config to allow the Docker setup too
localhost, 127.0.0.1 are not accessible from the outside, so bind to
`0.0.0.0` like we did with the Django server but set the *public* URL to
`localhost` because the server will read the manifest
file (webpack-stats.json) and use the URL for the browser to fetch the
asset so we do need a `localhost` that the browser can resolve.

Also remove some implicit BundleTracker options.