build: introduce the worker as a separate node
This makes the Docker environment feature-complete since no worker and
therefore no tasks were ran at all before this.

It also paves the way for that same separation to happen in production
because currently both the worker and the web server happen to run on
the same machine, which isn't the best idea for scalabilty or
resilience.

Introduce:
- a Redis node;
- a new "worker" service that runs off the tagged server
image ("ma-cantine-server");
- a Makefile to help administrate the Docker setup.